+++ src/requester.js
@@ -12,12 +12,8 @@ class Requester {
 
     return new Promise((resolve, reject) => {
       this.jsonp(
-        ((data) => {
-          resolve(data);
-        }),
-        (error) => {
-          reject(error);
-        }
+        ((data) => { resolve(data); }),
+        (error) => { reject(error); }
       );
     });
   }