+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### [0.0.3](https://github.com/jcep/jcep/releases/tag/v0.0.3)
+
+- Change endpoint correiosapi.apphb.com to viacep.com.br
+- Add complement attributes to API return
+
+### [0.0.2](https://github.com/jcep/jcep/releases/tag/v0.0.2)
+
+- Add beforeSend callback
+
 ### [0.0.1](https://github.com/jcep/jcep/releases/tag/v0.0.1)
 
 - The first release